Remove phf dependency from phf_mac, to simplify cross-compiling.
