+++ phf_mac/src/lib.rs
@@ -9,7 +9,6 @@
 extern crate rand;
 extern crate syntax;
 extern crate time;
-extern crate phf;
 extern crate rustc;
 
 use std::collections::HashMap;
@@ -26,6 +25,9 @@ use syntax::parse;
 use syntax::parse::token::{InternedString, COMMA, EOF, FAT_ARROW};
 use rand::{Rng, SeedableRng, XorShiftRng};
 use rustc::plugin::Registry;
+
+#[path="../../phf_shared/phf_shared.rs"]
+mod phf_shared;
 
 static DEFAULT_LAMBDA: uint = 5;
 
@@ -149,10 +151,10 @@ fn parse_map(cx: &mut ExtCtxt, tts: &[TokenTree]) -> Option<Vec<Entry>> {
         }
     }
 
-    if entries.len() > phf::MAX_SIZE {
+    if entries.len() > phf_shared::MAX_SIZE {
         cx.span_err(parser.span,
                     format!("maps with more than {} entries are not supported",
-                            phf::MAX_SIZE).as_slice());
+                            phf_shared::MAX_SIZE).as_slice());
         return None;
     }
 
@@ -189,10 +191,10 @@ fn parse_set(cx: &mut ExtCtxt, tts: &[TokenTree]) -> Option<Vec<Entry>> {
         }
     }
 
-    if entries.len() > phf::MAX_SIZE {
+    if entries.len() > phf_shared::MAX_SIZE {
         cx.span_err(parser.span,
                     format!("maps with more than {} entries are not supported",
-                            phf::MAX_SIZE).as_slice());
+                            phf_shared::MAX_SIZE).as_slice());
         return None;
     }
 
@@ -284,7 +286,7 @@ fn try_generate_hash(entries: &[Entry], rng: &mut XorShiftRng)
     let k2 = rng.gen();
 
     let hashes: Vec<Hashes> = entries.iter().map(|entry| {
-        let (g, f1, f2) = phf::hash(&entry.key_str.get(), k1, k2);
+        let (g, f1, f2) = phf_shared::hash(&entry.key_str.get(), k1, k2);
         Hashes {
             g: g,
             f1: f1,
@@ -312,7 +314,7 @@ fn try_generate_hash(entries: &[Entry], rng: &mut XorShiftRng)
             'disps_l: for d2 in range(0, table_len) {
                 try_map.clear();
                 for &key in bucket.keys.iter() {
-                    let idx = phf::displace(hashes.get(key).f1,
+                    let idx = phf_shared::displace(hashes.get(key).f1,
                                             hashes.get(key).f2,
                                             d1,
                                             d2) % table_len;
